work all

+++ src/main/java/com/faost/security/service/current/user/CurrentUserService.java
@@ -3,5 +3,5 @@ package com.faost.security.service.current.user;
 import com.faost.security.domain.security.CurrentUser;
 
 public interface CurrentUserService {
-    boolean canAccessUser(CurrentUser currentUser, Integer userId);
+    boolean canAccessUser(CurrentUser currentUser, Long userId);
 }
