+++ src/main/java/com/faost/security/service/current/user/CurrentUserServiceImpl.java
@@ -17,10 +17,10 @@ public class CurrentUserServiceImpl implements CurrentUserService {
     private RoleRepository roleRepository;
 
     @Override
-    public boolean canAccessUser(CurrentUser currentUser, Integer userId) {
+    public boolean canAccessUser(CurrentUser currentUser, Long userId) {
         LOGGER.debug("Checking if user={} has access to user={}", currentUser, userId);
         return currentUser != null
-                && (currentUser.getRole() == roleRepository.findOne(1L).getName()|| currentUser.getId().equals(userId));
+                && (currentUser.getRole() == roleRepository.findOne(1).getName()|| currentUser.getId().equals(userId));
     }
 
 }
